requirements.txt
pygame-ce
moviepy